Enabled everything that appeared disabled and made everything compile. Renamed kis_merge.h to kis_merge_visitor.h. Fixed the scaling (I hope, untested), but should do the rotate, shear and transform visitors. Made the histogram use paint devices. Added a KisThread class that sports a cancel boolean.
svn path=/trunk/koffice/; revision=492699

+++ krita/ui/kis_controlframe.cc
@@ -97,7 +97,7 @@ KisControlFrame::KisControlFrame( KMainWindow * /*window*/, KisView * view, cons
 
     KisConfig cfg;
     m_font  = KGlobalSettings::generalFont();
-    m_font.setPointSize(cfg.dockerFontSize());
+    m_font.setPointSize((int)cfg.dockerFontSize());
 
     m_brushWidget = new KisIconWidget(view, "brushes");
     m_brushWidget->setTextLabel( i18n("Brush Shapes") );
